+++ servidorrestspring/src/main/java/com/ipartek/formacion/spring/servidorrestspring/repositorios/Dao.java
@@ -2,7 +2,7 @@ package com.ipartek.formacion.spring.servidorrestspring.repositorios;
 
 public interface Dao<T> {
 	Iterable<T> obtenerTodos();
-	T obtenerPorId(Long id);
+	T obtenerPorId(int id);
 	
 	T insertar(T t);
 	T modificar(T t);
